+++ web/apps/wps-web/src/features/landingPage/components/FavouriteButton.tsx
