+++ web/apps/wps-web/src/features/landingPage/toolInfo.tsx
@@ -203,8 +203,7 @@ export const weatherToolkitInfo: ToolInfo = {
   managedBy: PREDICTIVE_SERVICES_UNIT
 }
 
-// The order of items in this array determines the order of items as they appear in the landing page
-// side bar and order of tiles.
+// The order of items in this array determines their order in the landing page sections and quick access menu.
 export const toolInfos = [
   fireBehaviourAdvisoryInfo,
   moreCastInfo,
